+++ code/techtest-application/src/main/java/com/bcnc/techtest/usecases/UpdateProductUseCaseImpl.java
@@ -6,6 +6,7 @@ import com.bcnc.techtest.domain.ports.out.ProductRepositoryPort;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Slf4j
@@ -18,7 +19,7 @@ public class UpdateProductUseCaseImpl implements UpdateProductUseCase {
     }
 
     @Override
-    public void updateProduct(long brandId, long productId, LocalDateTime date, Product product) {
+    public void updateProduct(BigDecimal brandId, BigDecimal productId, LocalDateTime date, Product product) {
         this.repository.updateProduct(brandId, productId, date, product);
         log.debug("|- Use Case: Update Product -| ");
         log.debug("|- Status: COMPLETE -|");
