+++ code/techtest-domain/src/main/java/com/bcnc/techtest/domain/ports/in/DeleteProductUseCase.java
@@ -1,7 +1,8 @@
 package com.bcnc.techtest.domain.ports.in;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface DeleteProductUseCase {
-    void deleteProduct(long brandId, long productId, LocalDateTime date);
+    void deleteProduct(BigDecimal brandId, BigDecimal productId, LocalDateTime date);
 }
