+++ code/techtest-domain/src/main/java/com/bcnc/techtest/domain/ports/in/UpdateProductUseCase.java
@@ -2,9 +2,10 @@ package com.bcnc.techtest.domain.ports.in;
 
 import com.bcnc.techtest.domain.models.Product;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface UpdateProductUseCase {
-    void updateProduct(long brandId, long productId, LocalDateTime date, Product product);
+    void updateProduct(BigDecimal brandId, BigDecimal productId, LocalDateTime date, Product product);
 }
