+++ code/techtest-infrastructure/techtest-data-jpa/src/main/java/com/bcnc/techtest/data/entities/ProductKey.java
@@ -3,9 +3,10 @@ package com.bcnc.techtest.data.entities;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 public class ProductKey implements Serializable {
-    private long productId;
-    private long brandId;
+    private BigDecimal productId;
+    private BigDecimal brandId;
 }
